refactor: decompose 1142-line frankenstein_editor.py into domain modules (fixes #2386)
