+++ src/tools/model_explorer/frankenstein_editor/__init__.py
@@ -1,0 +1,11 @@
+from .dialogs import StealComponentDialog
+from .editor import FrankensteinEditor
+from .model import URDFModel
+from .panel import ModelPanel
+
+__all__ = [
+    "FrankensteinEditor",
+    "URDFModel",
+    "ModelPanel",
+    "StealComponentDialog",
+]
